Unify camera and render tools toolbars into ViewToolbar
