+++ vibra/interface/plots/structural/displacement_field_inputs.py
@@ -27,7 +27,7 @@ class PlotDisplacementFieldInputs(DisplacementFieldInputs_UI):
         app().main_window.render_widget_changed.emit()
 
         app().main_window.animation_toolbar.setDisabled(False)
-        app().main_window.render_tools_toolbar.hide_selection_tool()
+        app().main_window.view_toolbar.hide_selection_tool()
 
     def _initialize(self):
         self.selected_frequency_index = None
